added putio debrid

+++ src/adapters/db.ts
@@ -16,9 +16,7 @@ class DB {
 			basepath = path.join(path.dirname(pkgjson.path), 'node_modules/.cache')
 		}
 		let file = path.join(basepath, `leveldb/${name}.db`)
-		if (process.DEVELOPMENT) {
-			fs.removeSync(file)
-		}
+		process.DEVELOPMENT && fs.removeSync(file)
 		fs.ensureDirSync(file)
 		return file
 	}
